Split pipeline tests into three suites
